+++ src/main/java/web/repository/UserRepository.java
@@ -7,11 +7,12 @@ import java.util.List;
 public interface UserRepository {
     void addUser(User user);
 
-    void deleteUser(Long id);
+    void deleteUser(int id);
 
     void editUser(User user);
 
-    User getUserById(Long id);
+    User getUserById(int id);
+
 
     List<User> getAllUsers();
 }
